chore: add warn logging to silent errors, extract shared test utilities
- Add tracing::warn! to 4 silent Err(_) => continue paths in staking.rs
  so operator data gaps are visible in logs
- Extract TestDb, get_db(), sample_operator(), sample_share_price() into
  test_utils.rs to eliminate duplication between staking and storage tests
- Remove unnecessary #[allow(dead_code)] on OperatorStatusCompact
- Discard unused InvalidBundle hash data during SCALE decode
